refactor : entity-module 패키지 명 변경
- entity-module.board -> question 으로 변경

연관 이슈
- #202
