+++ module-domain/src/main/java/com/checkping/domain/question/TaskBoardLink.java
@@ -1,4 +1,4 @@
-package com.checkping.domain.board;
+package com.checkping.domain.question;
 
 import com.checkping.domain.BaseEntity;
 import jakarta.persistence.Column;
@@ -10,28 +10,25 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.time.LocalDateTime;
+import java.util.Objects;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import org.springframework.data.annotation.LastModifiedDate;
 
 @Getter
 @ToString
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "task_board_partial_history")
+@Table(name = "task_board_link")
 @Entity
-public class TaskBoardPartialHistory extends BaseEntity {
+public class TaskBoardLink extends BaseEntity {
     /*
-    id : 업무 관리 게시글 부분이력 ID
-    updateName : 변경 항목 이름
-    currentValue : 현재 값
-    previousValue : 변경 이전 값
-    updatedAt : 수정 일시
+    id : 업무 관리 게시글 첨부 링크 아이디
+    name : 업무 관리 게시글 첨부 링크 이름
+    url : 업무 관리 게시글 첨부 링크 URL
     taskBoard : 업무 관리 게시글 (join)
      */
     @Id
@@ -39,20 +36,26 @@ public class TaskBoardPartialHistory extends BaseEntity {
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @Column(name = "update_name", nullable = false, length = 100)
-    private String updateName;
+    @Column(name = "name", nullable = false, length = 100)
+    private String name;
 
-    @Column(name = "current_value", nullable = false, length = 5000)
-    private String currentValue;
-
-    @Column(name = "previous_value", nullable = false, length = 5000)
-    private String previousValue;
-
-    @LastModifiedDate
-    @Column (name = "updated_at", nullable = false)
-    private LocalDateTime updatedAt;
+    @Column(name = "url", nullable = false, length = 2000)
+    private String url;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "task_board_id", nullable = false)
     private TaskBoard taskBoard;
+
+    @Override
+    public boolean equals(Object o) {
+        if (!(o instanceof TaskBoardLink that)) {
+            return false;
+        }
+        return Objects.equals(id, that.id) && Objects.equals(url, that.url);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, url);
+    }
 }
