+++ module-infra/src/main/java/com/checkping/infra/repository/project/taskboardcomment/TaskBoardCommentStoreImpl.java
@@ -1,6 +1,6 @@
 package com.checkping.infra.repository.project.taskboardcomment;
 
-import com.checkping.domain.board.TaskBoardComment;
+import com.checkping.domain.question.TaskBoardComment;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
