+++ module-infra/src/main/java/com/checkping/infra/repository/project/taskboardlink/TaskBoardLinkStore.java
@@ -1,6 +1,6 @@
 package com.checkping.infra.repository.project.taskboardlink;
 
-import com.checkping.domain.board.TaskBoardLink;
+import com.checkping.domain.question.TaskBoardLink;
 
 public interface TaskBoardLinkStore {
     TaskBoardLink store(TaskBoardLink taskBoardLink);
